[Refactor] Move ViewModel files to viewModel dir
